+++ src/extensions/intrahost/intrahost.jl
@@ -1,4 +1,4 @@
-function pathogenFractionsWinnerTakesAll(host::Host, population::Population)
+function pathogenFractionsWinnerTakesAll(host::Host, weightedResponse::FunctionWrapper{Float64,Tuple{Pathogen,Host,Int64}})
     # Currently, we assume pathogen population fraction (share in total fitness)
     # impacts all events equally regardless of event; this is not necessarilly the case, however.
     # We also assume only the most fit pathogen affects events.
@@ -10,7 +10,7 @@ function pathogenFractionsWinnerTakesAll(host::Host, population::Population)
         if length(host.responses) > 0
             for p in 1:length(host.pathogens)
                 fracs[p] =
-                    host.pathogens[p].coefficients[INTRAHOST_FITNESS] * population.parameters.weightedResponse(
+                    host.pathogens[p].coefficients[INTRAHOST_FITNESS] * weightedResponse(
                         host.pathogens[p], host, INTRAHOST_FITNESS
                     )
             end