+++ src/parameters/parameters.jl
@@ -10,46 +10,48 @@ const DEFAULT_PATHOGEN_TYPE = PathogenType(
     0.0,
     g->1.0, g->1.0, g->1.0, g->1.0,
     SA[ # order defined in COEFFICIENTS
-        g->1.0, g->1.0, g->1.0,
-        g->1.0, g->1.0, g->1.0,
-        g->1.0, g->1.0, g->1.0,
-        g->1.0, g->1.0, g->1.0
+        g::String->1.0, g::String->1.0, g::String->1.0,
+        g::String->1.0, g::String->1.0, g::String->1.0,
+        g::String->1.0, g::String->1.0, g::String->1.0,
+        g::String->1.0, g::String->1.0, g::String->1.0,
     ],
 )
 
 const DEFAULT_RESPONSE_TYPE = ResponseType(
     "Default",
     0.0,
-    (imp_g, mat_g, pat_g) -> 1.0,
-    (imp_g, mat_g, pat_g) -> 1.0,
+    (imp_g::String, mat_g::String, pat_g::String) -> 1.0,
+    (imp_g::String, mat_g::String, pat_g::String) -> 1.0,
     SA[ # order defined in COEFFICIENTS
-        (imp_g, mat_g) -> 1.0, (imp_g, mat_g) -> 1.0, (imp_g, mat_g) -> 1.0,
-        (imp_g, mat_g) -> 1.0, (imp_g, mat_g) -> 1.0, (imp_g, mat_g) -> 1.0,
-        (imp_g, mat_g) -> 1.0, (imp_g, mat_g) -> 1.0, (imp_g, mat_g) -> 1.0,
-        (imp_g, mat_g) -> 1.0, (imp_g, mat_g) -> 1.0, (imp_g, mat_g) -> 1.0,
+        (imp_g::String, mat_g::String) -> 1.0, (imp_g::String, mat_g::String) -> 1.0, (imp_g::String, mat_g::String) -> 1.0,
+        (imp_g::String, mat_g::String) -> 1.0, (imp_g::String, mat_g::String) -> 1.0, (imp_g::String, mat_g::String) -> 1.0,
+        (imp_g::String, mat_g::String) -> 1.0, (imp_g::String, mat_g::String) -> 1.0, (imp_g::String, mat_g::String) -> 1.0,
+        (imp_g::String, mat_g::String) -> 1.0, (imp_g::String, mat_g::String) -> 1.0, (imp_g::String, mat_g::String) -> 1.0,
     ],
     SA[ # order defined in COEFFICIENTS
-        (imp_g, mat_g, pat_g) -> 1.0, (imp_g, mat_g, pat_g) -> 1.0, (imp_g, mat_g, pat_g) -> 1.0,
-        (imp_g, mat_g, pat_g) -> 1.0, (imp_g, mat_g, pat_g) -> 1.0, (imp_g, mat_g, pat_g) -> 1.0,
-        (imp_g, mat_g, pat_g) -> 1.0, (imp_g, mat_g, pat_g) -> 1.0, (imp_g, mat_g, pat_g) -> 1.0,
-        (imp_g, mat_g, pat_g) -> 1.0, (imp_g, mat_g, pat_g) -> 1.0, (imp_g, mat_g, pat_g) -> 1.0
+        (imp_g::String, mat_g::String, pat_g::String) -> 1.0, (imp_g::String, mat_g::String, pat_g::String) -> 1.0,
+        (imp_g::String, mat_g::String, pat_g::String) -> 1.0, (imp_g::String, mat_g::String, pat_g::String) -> 1.0,
+        (imp_g::String, mat_g::String, pat_g::String) -> 1.0, (imp_g::String, mat_g::String, pat_g::String) -> 1.0,
+        (imp_g::String, mat_g::String, pat_g::String) -> 1.0, (imp_g::String, mat_g::String, pat_g::String) -> 1.0,
+        (imp_g::String, mat_g::String, pat_g::String) -> 1.0, (imp_g::String, mat_g::String, pat_g::String) -> 1.0,
+        (imp_g::String, mat_g::String, pat_g::String) -> 1.0, (imp_g::String, mat_g::String, pat_g::String) -> 1.0,
     ],
 )
 
 const DEFAULT_POPULATION_TYPE = PopulationType(
-    "Default", # id::String
+    "Default",
     true, false,
-    1.0, # inoculum_coefficient
-    1.0, # mutation_coefficient
-    1.0, # recombination_coefficient
-    SA[ # base_coefficients::SVector{NUM_COEFFICIENTS,Float64}
-        0.0, 1.0, 0.0,
+    1.0,
+    1.0,
+    1.0,
+    SA[ # order defined in COEFFICIENTS
+        0.0, 1.0,  0.0,
         0.0, 1.05, 0.0,
-        0.0, 0.0, 0.0,
-        0.0, 1.0, 0.0
+        0.0, 0.0,  0.0,
+        0.0, 1.0,  0.0,
     ],
     pathogenFractionsWinnerTakesAll,
     weightedResponseArithmeticMean,
     infectionProbabilityArithmeticMean,
-    (p, h, c) -> Nothing, # developResponses::Function,
+    (p::Pathogen, h::Host, r::Vector{Response}) -> Vector{Response}(undef,0),
 )